Added tests Reworked tests each on it's own file
