+++ MangaManager/CommonLib/HelperFunctions.py
@@ -1,5 +1,7 @@
 import time
 
+import requests
+
 
 def get_elapsed_time(start_time: float) -> str:
     """
@@ -45,6 +47,7 @@ def create_settings():
         "cover_folder_path": None
     }
 
+
 def cleanFilename(sourcestring, removestring="%:/,.\\[]<>*?\""):
     """Clean a string by removing selected characters.
 
@@ -66,3 +69,16 @@ def cleanFilename(sourcestring, removestring="%:/,.\\[]<>*?\""):
     # remove the undesireable characters
     new_string = ''.join([c for c in sourcestring if c not in removestring]).replace(" ", "_")
     return new_string
+
+
+def check_url_isImage(image_url):
+    """
+    Whether the url provided contains an image based on the headers
+    :param image_url:
+    :return:
+    """
+    image_formats = ("image/png", "image/jpeg", "image/jpg")
+    r = requests.head(image_url)
+    if r.headers["content-type"] in image_formats:
+        return True
+    return False
